# frontend/public/openapi.json
{
  "components": {
    "schemas": {
      "AcceptOfferInput": {
        "properties": {
          "issuedAt": {
            "type": "integer"
          },
          "offerId": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "issuedAt",
          "offerId",
          "signature"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "stable machine-readable code, e.g. invalid_request, unauthorized, replayed, chain_rejected",
                "type": "string"
              },
              "message": {
                "description": "human-readable explanation; never contains internal detail",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "FinancingOffer": {
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "feeBps": {
            "type": "integer"
          },
          "financier": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "accepted",
          "createdAt",
          "feeBps",
          "financier",
          "id"
        ],
        "type": "object"
      },
      "FinancingRequest": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "buyer": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "exporter": {
            "type": "string"
          },
          "externalRef": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "invoiceValue": {
            "type": "string"
          },
          "maxFeeBps": {
            "type": "integer"
          },
          "milestoneCount": {
            "type": "integer"
          },
          "note": {
            "type": "string"
          },
          "offers": {
            "items": {
              "$ref": "#/components/schemas/FinancingOffer"
            },
            "type": "array"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "pricing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Suggestion"
              },
              {
                "type": "null"
              }
            ]
          },
          "route": {
            "items": {
              "$ref": "#/components/schemas/RoutePoint"
            },
            "type": "array"
          },
          "shipmentId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "buyer",
          "createdAt",
          "exporter",
          "externalRef",
          "id",
          "invoiceValue",
          "maxFeeBps",
          "milestoneCount",
          "note",
          "offers",
          "policy",
          "route",
          "shipmentId",
          "status"
        ],
        "type": "object"
      },
      "FinancingRequestInput": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "issuedAt": {
            "type": "integer"
          },
          "maxFeeBps": {
            "type": "integer"
          },
          "milestoneCount": {
            "type": "integer"
          },
          "note": {
            "type": "string"
          },
          "shipmentId": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "amount",
          "issuedAt",
          "maxFeeBps",
          "milestoneCount",
          "shipmentId",
          "signature"
        ],
        "type": "object"
      },
...
      "OfferInput": {
        "properties": {
          "address": {
            "type": "string"
          },
          "feeBps": {
            "type": "integer"
          },
          "issuedAt": {
            "type": "integer"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "feeBps",
          "issuedAt",
          "signature"
        ],
        "type": "object"
      },
...
      "Policy": {
        "properties": {
          "maxConflictBps": {
            "type": "integer"
          },
          "maxGapSec": {
            "type": "integer"
          },
          "maxHumidityX100": {
            "type": "integer"
          },
          "maxRiskBps": {
            "type": "integer"
          },
          "maxRouteDeviationM": {
            "type": "integer"
          },
          "maxShockX100": {
            "type": "integer"
          },
          "maxTempX100": {
            "type": "integer"
          },
          "minEvidenceScore": {
            "type": "integer"
          },
          "minSensors": {
            "type": "integer"
          },
          "minTempX100": {
            "type": "integer"
          },
          "requiresZk": {
            "type": "boolean"
          }
        },
        "required": [
          "maxConflictBps",
          "maxGapSec",
          "maxHumidityX100",
          "maxRiskBps",
          "maxRouteDeviationM",
          "maxShockX100",
          "maxTempX100",
          "minEvidenceScore",
          "minSensors",
          "minTempX100",
          "requiresZk"
        ],
        "type": "object"
      },
...
      "RequestList": {
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/FinancingRequest"
            },
            "type": "array"
          }
        },
        "required": [
          "requests"
        ],
        "type": "object"
      },
...
      "ShipmentView": {
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Cover"
              },
              {
                "type": "null"
              }
            ]
          },
          "facility": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FacilityView"
              },
              {
                "type": "null"
              }
            ]
          },
          "latestEvidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EpochSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "milestones": {
            "items": {
              "$ref": "#/components/schemas/Milestone"
            },
            "type": "array"
          },
          "openCoverOffers": {
            "type": "integer"
          },
          "quarantinedReadings": {
            "type": "integer"
          },
          "shipment": {
            "$ref": "#/components/schemas/Shipment"
          },
          "title": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TitleView"
              },
              {
                "type": "null"
              }
            ]
          },
          "usdgDecimals": {
            "type": "integer"
          }
        },
        "required": [
          "cover",
          "facility",
          "latestEvidence",
          "milestones",
          "openCoverOffers",
          "quarantinedReadings",
          "shipment",
          "title",
          "usdgDecimals"
        ],
        "type": "object"
      },
      "SignedAuthorization": {
        "properties": {
          "issuedAt": {
            "type": "integer"
          },
          "signature": {
            "type": "string"
          }
        },
        "required": [
          "issuedAt",
          "signature"
        ],
        "type": "object"
      },
...
      "Suggestion": {
        "properties": {
          "highBps": {
            "type": "integer"
          },
          "inputs": {
            "$ref": "#/components/schemas/Inputs"
          },
          "lowBps": {
            "type": "integer"
          },
          "midBps": {
            "type": "integer"
          },
          "model": {
            "description": "model identifier; the formula is documented in the API reference and README",
            "type": "string"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/Reason"
            },
            "type": "array"
          },
          "shipmentId": {
            "type": "string"
          }
        },
        "required": [
          "highBps",
          "inputs",
          "lowBps",
          "midBps",
          "model",
          "reasons",
          "shipmentId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminKey": {
        "description": "the operator's ADMIN_API_KEY",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "alchemySignature": {
        "description": "Alchemy Notify: hex HMAC-SHA256 of the raw request body with the webhook's signing key (ALCHEMY_WEBHOOK_SIGNING_KEYS)",
        "in": "header",
        "name": "X-Alchemy-Signature",
        "type": "apiKey"
      },
      "sourceSignature": {
        "description": "Signed by a registered evidence source. The signing string is\n\n```\nCARGOFLOW-V1\\n<METHOD>\\n<PATH>\\n<TIMESTAMP>\\n<hex sha256(body)>\n```\n\n`ed25519` sources sign it with Ed25519. `p256` sources sign it with ECDSA P-256 over SHA-256 (X-Signature is the DER or raw 64-byte r||s signature). `webauthn` sources run navigator.credentials.get with challenge = sha256(signing string) (32 raw bytes) and send the assertion: X-Signature = signature, X-WebAuthn-Authenticator-Data = authenticatorData, X-WebAuthn-Client-Data = clientDataJSON (all base64url); the backend checks type `webauthn.get`, the challenge, the registered RP id hash, the user-present flag and the signature over authenticatorData || sha256(clientDataJSON).",
        "in": "header",
        "name": "X-Signature",
        "type": "apiKey"
      },
      "walletSignature": {
        "description": "Not a header: the body carries `issuedAt` and `signature` (hex EIP-191 personal_sign) over the operation's `x-cargoflow-signed-message`. Declared as a scheme so clients can see which operations need a wallet.",
        "in": "header",
        "name": "X-Wallet-Signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "CargoFlow turns signed shipment telemetry into on-chain financing decisions. Public reads need no credentials. Writes are authenticated in one of three ways: the operator's admin key, a party's wallet signature over a fixed message (see each operation's `x-cargoflow-signed-message`), or an evidence source's request signature (Ed25519, P-256 or WebAuthn). Errors are `{\"error\": {\"code\", \"message\"}}` with stable codes. Amounts are USDG base units (6 decimals) as decimal strings.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "CargoFlow API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/requests": {
      "get": {
        "description": "Each request carries `pricing`, the fee guidance of GET /v1/pricing/suggest.",
        "operationId": "listRequests",
        "parameters": [
          {
            "description": "open, accepted, funded or closed",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "a 0x address",
            "in": "query",
            "name": "exporter",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestList"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "Financing requests, newest first, with offers cheapest first",
        "tags": [
          "Marketplace"
        ]
      },
      "post": {
        "description": "Signed by a wallet (EIP-191 personal_sign, or EIP-1271 for contract wallets) over this exact text, lines joined by \\n with no trailing newline, ids and addresses lower case; `issued` is unix seconds within 10 minutes of the server clock, and each authorization is single-use (409 `replayed`):\n\n```\nCargoFlow financing request\nshipment: <id>\namount: <base units>\nmax fee bps: <n>\nmilestones: <n>\nissued: <t>\n```",
        "operationId": "createRequest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancingRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancingRequest"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "walletSignature": []
          }
        ],
        "summary": "Post a financing request (exporter)",
        "tags": [
          "Marketplace"
        ],
        "x-cargoflow-signed-message": "CargoFlow financing request\nshipment: <id>\namount: <base units>\nmax fee bps: <n>\nmilestones: <n>\nissued: <t>"
      }
    },
    "/v1/requests/{rid}/accept": {
      "post": {
        "description": "Signed by a wallet (EIP-191 personal_sign, or EIP-1271 for contract wallets) over this exact text, lines joined by \\n with no trailing newline, ids and addresses lower case; `issued` is unix seconds within 10 minutes of the server clock, and each authorization is single-use (409 `replayed`):\n\n```\nCargoFlow accept\nrequest: <rid>\noffer: <offer id>\nissued: <t>\n```",
        "operationId": "acceptOffer",
        "parameters": [
          {
            "description": "financing request id (uuid)",
            "in": "path",
            "name": "rid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptOfferInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancingRequest"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "walletSignature": []
          }
        ],
        "summary": "Accept an offer (exporter)",
        "tags": [
          "Marketplace"
        ],
        "x-cargoflow-signed-message": "CargoFlow accept\nrequest: <rid>\noffer: <offer id>\nissued: <t>"
      }
    },
    "/v1/requests/{rid}/close": {
      "post": {
        "description": "Signed by a wallet (EIP-191 personal_sign, or EIP-1271 for contract wallets) over this exact text, lines joined by \\n with no trailing newline, ids and addresses lower case; `issued` is unix seconds within 10 minutes of the server clock, and each authorization is single-use (409 `replayed`):\n\n```\nCargoFlow close request\nrequest: <rid>\nissued: <t>\n```",
        "operationId": "closeRequest",
        "parameters": [
          {
            "description": "financing request id (uuid)",
            "in": "path",
            "name": "rid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignedAuthorization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancingRequest"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "walletSignature": []
          }
        ],
        "summary": "Withdraw a request (exporter)",
        "tags": [
          "Marketplace"
        ],
        "x-cargoflow-signed-message": "CargoFlow close request\nrequest: <rid>\nissued: <t>"
      }
    },
    "/v1/requests/{rid}/offers": {
      "post": {
        "description": "Signed by a wallet (EIP-191 personal_sign, or EIP-1271 for contract wallets) over this exact text, lines joined by \\n with no trailing newline, ids and addresses lower case; `issued` is unix seconds within 10 minutes of the server clock, and each authorization is single-use (409 `replayed`):\n\n```\nCargoFlow offer\nrequest: <rid>\nfee bps: <n>\nissued: <t>\n```",
        "operationId": "placeOffer",
        "parameters": [
          {
            "description": "financing request id (uuid)",
            "in": "path",
            "name": "rid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "replaced your earlier fee"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancingOffer"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [
          {
            "walletSignature": []
          }
        ],
        "summary": "Offer a fee (a financier)",
        "tags": [
          "Marketplace"
        ],
        "x-cargoflow-signed-message": "CargoFlow offer\nrequest: <rid>\nfee bps: <n>\nissued: <t>"
      }
    },
...
    "/v1/shipments/{id}": {
      "get": {
        "operationId": "getShipment",
        "parameters": [
          {
            "description": "shipment id: 0x followed by 64 lowercase hex characters (keccak256(exporter, keccak256(externalRef)))",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipmentView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "error"
          }
        },
        "security": [],
        "summary": "Combined store and live chain view of a shipment",
        "tags": [
          "Shipments"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "this backend",
      "url": "/"
    }
  ]
}